post_updater/requirements.txt
requests==2.21.0
google_api_python_client==1.12.8
python_dateutil==2.8.1
PyYAML==5.4.1
